update label method and test
