+++ pkg/label/label.go
@@ -14,27 +14,20 @@ const (
 	ManagedBy = "giantswarm.io/managed-by"
 )
 
-// ProcessLabels ensures the chart-operator.giantswarm.io/version label is
-// present and the app-operator.giantswarm.io/version label is removed. It
-// also ensures the giantswarm.io/managed-by label is accurate.
-//
-// Any other labels added to the app custom resource are passed on to the chart
-// custom resource.
-func ProcessLabels(projectName, chartCustomResourceVersion string, inputLabels map[string]string) map[string]string {
-	// These labels are required.
-	labels := map[string]string{
-		ManagedBy: projectName,
-	}
-
-	if chartCustomResourceVersion != "" {
-		labels[ChartOperatorVersion] = chartCustomResourceVersion
-	}
+// ProcessLabels merge the requiredLabels into inputLabels, also remove all excludedLabels elements.
+func ProcessLabels(inputLabels, requiredLabels, excludedLabels map[string]string) map[string]string {
+	labels := map[string]string{}
 
 	for k, v := range inputLabels {
 		// These labels must be removed.
-		if k != ManagedBy && k != AppOperatorVersion {
+		if _, ok := excludedLabels[k]; !ok {
 			labels[k] = v
 		}
+	}
+
+	// These labels must be included.
+	for k, v := range requiredLabels {
+		labels[k] = v
 	}
 
 	return labels
