+++ service/controller/app/v1/resource/chart/desired_test.go
@@ -11,7 +11,6 @@ import (
 	"github.com/google/go-cmp/cmp"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/giantswarm/app-operator/pkg/label"
 	"github.com/giantswarm/app-operator/service/controller/app/v1/controllercontext"
 	"github.com/giantswarm/app-operator/service/controller/app/v1/key"
 )
@@ -213,66 +212,6 @@ func Test_Resource_GetDesiredState(t *testing.T) {
 				if !reflect.DeepEqual(chart.TypeMeta, tc.expectedChart.TypeMeta) {
 					t.Fatalf("want matching typemeta \n %s", cmp.Diff(chart.TypeMeta, tc.expectedChart.TypeMeta))
 				}
-			}
-		})
-	}
-}
-
-func Test_processLabels(t *testing.T) {
-	tests := []struct {
-		name           string
-		projectName    string
-		chartVersion   string
-		inputLabels    map[string]string
-		expectedLabels map[string]string
-	}{
-		{
-			name:         "case 0: basic match",
-			projectName:  "app-operator",
-			chartVersion: "1.0.0",
-			inputLabels: map[string]string{
-				"giantswarm.io/managed-by": "release-operator",
-			},
-			expectedLabels: map[string]string{
-				"chart-operator.giantswarm.io/version": "1.0.0",
-				"giantswarm.io/managed-by":             "app-operator",
-			},
-		},
-		{
-			name:         "case 1: extra labels still present",
-			projectName:  "app-operator",
-			chartVersion: "1.0.0",
-			inputLabels: map[string]string{
-				"app":                        "prometheus",
-				"giantswarm.io/cluster":      "5xchu",
-				"giantswarm.io/managed-by":   "cluster-operator",
-				"giantswarm.io/organization": "giantswarm",
-			},
-			expectedLabels: map[string]string{
-				"app":                                  "prometheus",
-				"chart-operator.giantswarm.io/version": "1.0.0",
-				"giantswarm.io/cluster":                "5xchu",
-				"giantswarm.io/managed-by":             "app-operator",
-				"giantswarm.io/organization":           "giantswarm",
-			},
-		},
-		{
-			name:         "case 2: empty inputs",
-			projectName:  "app-operator",
-			chartVersion: "",
-			expectedLabels: map[string]string{
-				"giantswarm.io/managed-by": "app-operator",
-			},
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-
-			result := label.ProcessLabels(tc.projectName, tc.chartVersion, tc.inputLabels)
-
-			if !reflect.DeepEqual(result, tc.expectedLabels) {
-				t.Fatalf("want matching \n %s", cmp.Diff(result, tc.expectedLabels))
 			}
 		})
 	}
